Add comments and split large file
