+++ tsify-next-macros/src/container.rs
@@ -1,12 +1,19 @@
 use serde_derive_internals::{ast, ast::Container as SerdeContainer, attr};
 
-use crate::{attrs::TsifyContainerAttrs, ctxt::Ctxt};
+use crate::{attrs::TsifyContainerAttrs, error_tracker::ErrorTracker};
 
+/// Data structure storing information about a type decorated with `#[derive(Tsify)]`.
+/// This structure also keeps information that was parsed via Serde's macros.
 pub struct Container<'a> {
-    pub ctxt: Ctxt,
+    /// Errors that occurred during processing.
+    pub errors: ErrorTracker,
+    /// Attributes passed to the `#[derive(Tsify)]` macro.
     pub attrs: TsifyContainerAttrs,
+    /// Information about the type as parsed by Serde.
     pub serde_container: SerdeContainer<'a>,
+    /// The `ident` of the type as written in the Rust code.
     pub ident_str: String,
+    /// The name type that will be serialized to Typescript.
     pub name: String,
 }
 
@@ -14,12 +21,12 @@ impl<'a> Container<'a> {
     pub fn new(serde_container: SerdeContainer<'a>) -> Self {
         let input = &serde_container.original;
         let attrs = TsifyContainerAttrs::from_derive_input(input);
-        let ctxt = Ctxt::new();
+        let errors = ErrorTracker::new();
 
         let attrs = match attrs {
             Ok(attrs) => attrs,
             Err(err) => {
-                ctxt.syn_error(err);
+                errors.syn_error(err);
                 Default::default()
             }
         };
@@ -33,7 +40,7 @@ impl<'a> Container<'a> {
             .format_name(serde_container.ident.to_string());
 
         Self {
-            ctxt,
+            errors,
             attrs,
             serde_container,
             ident_str,
@@ -55,10 +62,12 @@ impl<'a> Container<'a> {
         }
     }
 
+    /// The `ident` of the type as written in the Rust code.
     pub fn ident(&self) -> &syn::Ident {
         &self.serde_container.ident
     }
 
+    /// The `ident` of the type as written in the Rust code as a string.
     pub fn ident_str(&self) -> String {
         self.ident_str.clone()
     }
@@ -68,27 +77,33 @@ impl<'a> Container<'a> {
         &self.serde_container.attrs
     }
 
+    /// Whether or not Serde has marked this type as `transparent`.
     pub fn transparent(&self) -> bool {
         self.serde_attrs().transparent()
     }
 
+    /// The name of the type that will be serialized to Typescript.
     pub fn name(&self) -> String {
         self.name.clone()
     }
 
+    /// Information about the generics associated with the type as parsed by Serde.
     pub fn generics(&self) -> &syn::Generics {
         self.serde_container.generics
     }
 
+    /// Information about the data fields of the type as parsed by Serde.
     pub fn serde_data(&self) -> &ast::Data {
         &self.serde_container.data
     }
 
+    /// Add a new error to the list of processing errors.
     pub fn syn_error(&self, err: syn::Error) {
-        self.ctxt.syn_error(err);
+        self.errors.syn_error(err);
     }
 
+    /// Return all accumulated errors.
     pub fn check(self) -> syn::Result<()> {
-        self.ctxt.check()
+        self.errors.check()
     }
 }
